MAMA: Add the capability to load payload libraries just-in-time (JIT).
If a library hasn't been previously loaded then OpenMAMA will now attempt
to load it by converting the payload identifier into a well-known string
for the payload library.  This behaviour is useful if the payload isn't
known ahead of time or if lazy loading is desired over enumeration of a
existing set of libraries.  No changes made to existing API interfaces.

The following functions have been added :-

- mamaPayload_convertToLibString():
  Maps a mamaPayloadType enumeration to the well-known library name.

- mamaPayload_convertFromLibString():
  Inverse of mamaPayload_convertToLibString.

- mama_loadPayloadBridgeByType():
  Takes a mamaPayloadType and utilises mamaPayload_convertToLibString()
  to retrieve the library name for the payload, and then delegates this
  to the existing mama_loadPayloadBridge() function.

The string conversions are implemented as a linear search, primarily due
to these functions not being on the hot path - Their use is prior to a
library being loaded and (I would imagine) would be seldom used in other
situations.

Additional changes :-

If msgPayload_destroyImpl has been exposed by the payload library then
this is now utilised to cleanup the payload upon shutdown properly.  It
was observed that each payload always leaked memory due to lacking the
capability to destroy itself.  Individual payloads will still need to
expose this - Backwards compatible as it won't be called otherwise.

The MamaEnumTestsC set of unit tests have been rewritten to utilise the
new conversion functionality.  The tests have also been renamed as such:

- testPayloadToString -> testPayloadStringConversions
- testPayloadFromString -> testPayloadStringConversions
- testMiddlewareToString -> testMiddlewareStringConversions
- testMiddlewareFromString -> testMiddlewareStringConversions

Noticed that the ASSERT_EQ/EXPECT_EQ statements in gtest are sometimes
used incorrectly, in that the expected/actual are reversed (it should be
expected on the left and actual on the right).

Todo :-

Need to confirm well-known library names with all vendors - Will only be
important if JIT loading is actually used, but the names should be made
accurate anyway.